Add files via upload
Files for the fifth lab in a course on numerical computing I gave at University College Cork. Some files from the fourth lab are re-used. Lab introduces classes and how they are used to solve numerical problems, specifically, an object for finding the roots of a one-dimensional function is implemented. The STL container <vector> is also introduced.